Добавит слайс userSlice
