+++ src/types/state.ts
@@ -1,6 +1,29 @@
 import { store } from '../store';
+import { CityKeys } from './cities';
+import { OfferType, ShortOfferListType } from './offers';
+import { ReviewsListType } from './review';
+import { SortTypeKeys } from './types';
+import { AuthorizationStatusKeys, UserInfo } from './user';
 
-export type State = ReturnType<typeof store.getState>;
+type InitialState = {
+  currentCity: CityKeys;
+  currentSortKey: SortTypeKeys;
+  offers: ShortOfferListType;
+  isOffersLoading: boolean;
+  authorizationStatus: AuthorizationStatusKeys;
+  userInfo: UserInfo | null;
+  extendedOffer: OfferType | null;
+  isExtendedOfferLoading: boolean;
+  nearbyOffers: ShortOfferListType;
+  isNearbyOffersLoading: boolean;
+  reviewsList: ReviewsListType;
+  isReviewsListLoading: boolean;
+  isSubmitReviewLoading: boolean;
+  error: string | null;
+};
 
-export type AppDispatch = typeof store.dispatch;
+type State = ReturnType<typeof store.getState>;
 
+type AppDispatch = typeof store.dispatch;
+
+export type { InitialState, State, AppDispatch };
